backend/requirements.txt
# ===== CORE FRAMEWORK =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ===== DATABASE & ORM =====
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9       # PostgreSQL adapter
asyncpg==0.29.0              # Async PostgreSQL adapter

# ===== AUTHENTICATION & SECURITY =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-decouple==3.8
cryptography==41.0.7

# ===== DATA VALIDATION =====
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ===== CACHING & BACKGROUND TASKS =====
redis==5.0.1
celery==5.3.4
flower==2.0.1                # Celery monitoring

# ===== DATE & TIME HANDLING =====
python-dateutil==2.8.2

# ===== FILE HANDLING =====
aiofiles==23.2.1
pillow==10.1.0               # Image processing

# ===== API DOCUMENTATION =====
python-multipart==0.0.6

# ===== EXTERNAL INTEGRATIONS =====
requests==2.31.0             # HTTP requests for external APIs
httpx==0.25.2                # Async HTTP client

# ===== SMS & EMAIL =====
twilio==8.10.2               # SMS gateway
sendgrid==6.10.0             # Email service
python-emails==0.6

# ===== TESTING (Development) =====
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# ===== DEVELOPMENT TOOLS =====
black==23.11.0               # Code formatting
isort==5.12.0                # Import sorting
flake8==6.1.0                # Linting
mypy==1.7.1                  # Type checking

# ===== MONITORING & LOGGING =====
sentry-sdk[fastapi]==1.38.0  # Error tracking
structlog==23.2.0            # Structured logging

# ===== EXPORT & REPORTING =====
openpyxl==3.1.2              # Excel export
reportlab==4.0.7             # PDF generation
matplotlib==3.8.2           # Charts for reports
pandas==2.1.3                # Data analysis

# ===== M-PESA INTEGRATION =====
daraja==1.0.0                # M-Pesa SDK (if available)
# OR build custom M-Pesa integration